pyproject.toml
[project]
name = "vaudio-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.63.2",
    "python-dotenv>=1.0.1",
    "streamlit>=1.42.2",
]
